backend/requirements.txt
django-filter
djangorestframework-simplejwt
asgiref==3.9.1
channels==4.3.1
Django==5.2.5
djangorestframework==3.16.1
sqlparse==0.5.3
typing_extensions==4.14.1
